# SRC/types/responses/admin.responses.ts
export interface SaveUpdateResADMIN {
    _id: string;
    FirstName: string;
    LastName: string;
    Desgination: string;
    cell: string;
    JoinDate: string;
    adress: string;
    createdAt: string;
    updatedAt: string;
}